feat: add run context
This will allow running commands with the same context.
Could be useful to run multiple commands,
for example when building & flashing.
